+++ src/Plutarch/Extra/TermCont.hs
@@ -4,6 +4,8 @@
 module Plutarch.Extra.TermCont (
     module Extra,
     pguardWithC,
+    pguardShowC,
+    pexpectJustC,
 ) where
 
 import "plutarch-extra" Plutarch.Extra.TermCont as Extra (
@@ -16,6 +18,7 @@ import "plutarch-extra" Plutarch.Extra.TermCont as Extra (
     ptryFromC,
  )
 import Plutarch.Prelude
+import Plutarch.Show (PShow)
 
 {- | 'pguardC' but with type threading for better traces.
 
@@ -28,7 +31,7 @@ import Plutarch.Prelude
   This is great, but won't tell us what @foo@ _is_, when it isn't even.
   Thankfully, we can augment this using 'pguardWithC':
 
-  @'pguardWithC' (\x -> "foo should be even. it was " <> pshow x) ('peven' #) foo@
+  @'pguardWithC' (\x -> "foo should be even. It was " <> pshow x) ('peven' #) foo@
 
   @since 1.1.0
 -}
@@ -43,3 +46,32 @@ pguardWithC ::
     TermCont @r s ()
 pguardWithC tracer checker object =
     pguardC (tracer object) (checker object)
+
+{- | 'pguardWithC' but always uses 'PShow' instance to
+     generate trace result. Appends to assertion message.
+
+  @since 1.1.0
+-}
+pguardShowC ::
+    forall (r :: PType) (pt :: PType) (s :: S).
+    PShow pt =>
+    Term s PString ->
+    (Term s pt -> Term s PBool) ->
+    Term s pt ->
+    TermCont @r s ()
+pguardShowC message =
+    pguardWithC (\t -> message <> " Guarded object was: " <> pshow t)
+
+{- | Escape with a particular value on expecting 'Just'. For use in monadic context.
+
+    @since 1.1.0
+-}
+pexpectJustC ::
+    forall (a :: S -> Type) (r :: S -> Type) (s :: S).
+    Term s r ->
+    Term s (PMaybe a) ->
+    TermCont @r s (Term s a)
+pexpectJustC escape ma = tcont $ \f ->
+    pmatch ma $ \case
+        PJust v -> f v
+        PNothing -> escape